requirements.txt
boto3==1.34.162
botocore==1.34.162
pillow==10.4.0
numpy==1.26.4
